{{cookiecutter.project_name}}/requirements/common.txt
asgiref==3.4.1
Django==3.2.7
django-rest-framework==0.1.0
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
PyJWT==1.7.1

pytz==2021.1
sqlparse==0.4.1

{%- if cookiecutter.db_choice == "postgresql" %}
mysqlclient==2.0.3
PyMySQL==1.0.2
psycopg2
{%- endif %}

{%- if cookiecutter.cache_choice == "redis" %}
django-redis==5.0.0
redis==3.5.3
python-redis-lock==3.7.0
{%- endif %}
{%- if cookiecutter.use_simpleui == "yes" %}
django-simpleui
{%- endif %}
{%- if cookiecutter.use_ldap == "yes" %}
django-auth-ldap==3.0.0
ldap3
{%- endif %}
pillow

